+++ ft_philo/.gitignore
@@ -1,3 +1,5 @@
 test*.c
 **/*.o
 **/*.out
+**/.vscode/
+philo/philo
