--- ft_philo/philo/.vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "files.associations": {
-        "ft_philo.h": "c"
-    }
-}
