+++ ft_philo/philo/ft_errors.c
@@ -1,15 +1,3 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_errors.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ael-moha <ael-moha@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/05 21:13:10 by ael-moha          #+#    #+#             */
-/*   Updated: 2024/11/07 18:30:08 by ael-moha         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
 #include "ft_philo.h"
 
 void	f_mtx(void *ptr)
@@ -20,13 +8,15 @@ void	f_mtx(void *ptr)
 
 void	ft_free_all(t_all *all)
 {
-	ft_destroy_mutexes(all);
-	free(all->forks);
-	free(all->threads);
-	free(all->philos);
-	free(all->mutexes);
-	if (all->eat_count > 0)
-		free(all->mutex_eat_counter);
+	// ft_destroy_mutexes(all);
+	// free(all->forks);
+	// free(all->threads);
+	// free(all->philos);
+	// free(all->output_mtx);
+	// free(all->meal_mtx);
+	// if (all->eat_count > 0)
+	// 	free(all->mutex_eat_counter);
+	                   // FIX THE SGV HERE ITS NOT A BFD
 }
 
 void	ft_perror(char c)
@@ -43,20 +33,21 @@ void	ft_perror(char c)
 
 int	ft_check_simulation(t_philo *philo)
 {
-	int	res;
-
-	res = 1;
 	LOCK(philo->all->dead_lock);
 	if (!philo->all->simulation_running)
-		res = 0;
+		return (UNLOCK(philo->all->dead_lock), false);
 	UNLOCK(philo->all->dead_lock);
-	return (res);
+	return (true);
 }
 
 void	declare_death(t_philo *philo)
 {
 	LOCK(philo->all->dead_lock);
-	ft_p_action('d', philo);
-	philo->all->simulation_running = 0;
+		philo->all->simulation_running = 0;
 	UNLOCK(philo->all->dead_lock);
+
+	LOCK(philo->all->output_mtx);
+	printf("%zu %d died\n", current_time_in_milliseconds()
+		- philo->all->start_time, philo->index);
+	UNLOCK(philo->all->output_mtx);
 }
