+++ ft_philo/philo/ft_utils2.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_utils2.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ael-moha <ael-moha@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ysbai-jo <ysbai-jo@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 21:12:55 by ael-moha          #+#    #+#             */
-/*   Updated: 2024/11/23 02:46:46 by ael-moha         ###   ########.fr       */
+/*   Updated: 2024/11/23 15:41:36 by ysbai-jo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,11 +48,11 @@ int	ft_atoi(const char *nbr)
 	return (res * sign);
 }
 
-void	ft_p_action(char m, t_philo *philo)
+bool	ft_p_action(char m, t_philo *philo)
 {
-	// if (!ft_check_simulation(philo))
-	// 	return ;
 	LOCK(philo->all->output_mtx);
+	if (!ft_check_simulation(philo))
+		return (UNLOCK(philo->all->output_mtx), false);
 	if (m == 'e')
 		printf(AC_RED "%zu %d is eating\n" RESET, current_time_in_milliseconds()
 			- philo->all->start_time, philo->index);
@@ -64,8 +64,6 @@ void	ft_p_action(char m, t_philo *philo)
 		printf(AC_BLUE "%zu %d is thinking\n" RESET,
 			current_time_in_milliseconds() - philo->all->start_time,
 			philo->index);
-	else if (m == 'd')
-		printf("%zu %d died\n", current_time_in_milliseconds()
-			- philo->all->start_time, philo->index);
 	UNLOCK(philo->all->output_mtx);
+	return (true);
 }
